add hovers to text and sidebar buttons
