+++ src/components/UIsettings.css
@@ -96,7 +96,7 @@
 /* Ensure visibility and restore padding when active */
 .custom-settings-container.visible {
     opacity: 1;
-    max-height: 580px;
+    max-height: 800px;
     padding: 15px; /* Restore padding */
 }
 
@@ -133,7 +133,6 @@
     border-radius: 8px; /* Slightly more rounded corners */
     box-shadow: 0 5px 9px rgba(0, 0, 0, 0.15); /* More noticeable shadow */
     
-
 }
 
 .custom-color-selector {
@@ -142,6 +141,7 @@
     border: 1px !important;
     background-color: #e0e0e0;
     cursor: pointer;
+    padding: 0;
 }
 
 .button.btn.btn-primary{
