+++ src/index.css
@@ -5,7 +5,7 @@
 
 body {
   margin: 0;
-  font-family: "Inter", sans-serif !important; 
+  font-family: "Inter", sans-serif; 
   font-optical-sizing: auto !important;
   font-weight: 400 !important;
   font-style: normal !important;
